refactor: decompose parser package — split oversized files into focused modules
