+++ pkg/sql/parser/expressions.go
@@ -13,7 +13,11 @@
 // limitations under the License.
 
 // Package parser - expressions.go
-// Expression parsing functions for the SQL parser.
+// Core expression parsing: precedence-climbing hierarchy from OR down to JSON/cast operators.
+// Related modules:
+//   - expressions_literal.go   — literals, identifiers, qualified names, parsePrimaryExpression
+//   - expressions_operators.go — comparison, BETWEEN, IN, LIKE, IS NULL, array subscripts
+//   - expressions_complex.go   — CASE, CAST, INTERVAL, ARRAY constructors
 
 package parser
 
@@ -24,7 +28,6 @@ import (
 	goerrors "github.com/ajitpratap0/GoSQLX/pkg/errors"
 	"github.com/ajitpratap0/GoSQLX/pkg/models"
 	"github.com/ajitpratap0/GoSQLX/pkg/sql/ast"
-	"github.com/ajitpratap0/GoSQLX/pkg/sql/keywords"
 )
 
 // parseExpression parses an expression with OR operators (lowest precedence)
@@ -103,290 +106,6 @@ func (p *Parser) parseAndExpression() (ast.Expression, error) {
 			Right:    right,
 			Pos:      opPos,
 		}
-	}
-
-	return left, nil
-}
-
-// parseComparisonExpression parses an expression with comparison operators
-func (p *Parser) parseComparisonExpression() (ast.Expression, error) {
-	// Parse the left side using string concatenation expression for arithmetic support
-	left, err := p.parseStringConcatExpression()
-	if err != nil {
-		return nil, err
-	}
-
-	// Check for NOT prefix for BETWEEN, LIKE, IN operators
-	// Only consume NOT if followed by BETWEEN, LIKE, ILIKE, or IN
-	// This prevents breaking cases like: WHERE NOT active AND name LIKE '%'
-	notPrefix := false
-	if p.isType(models.TokenTypeNot) {
-		nextToken := p.peekToken()
-		nextUpper := strings.ToUpper(nextToken.Token.Value)
-		if nextUpper == "BETWEEN" || nextUpper == "LIKE" || nextUpper == "ILIKE" || nextUpper == "IN" {
-			notPrefix = true
-			p.advance() // Consume NOT only if followed by valid operator
-		}
-	}
-
-	// Check for BETWEEN operator
-	if p.isType(models.TokenTypeBetween) {
-		betweenPos := p.currentLocation()
-		p.advance() // Consume BETWEEN
-
-		// Parse lower bound - use parseStringConcatExpression to support complex expressions
-		// like: price BETWEEN price * 0.9 AND price * 1.1
-		lower, err := p.parseStringConcatExpression()
-		if err != nil {
-			return nil, goerrors.InvalidSyntaxError(
-				fmt.Sprintf("failed to parse BETWEEN lower bound: %v", err),
-				p.currentLocation(),
-				p.currentToken.Token.Value,
-			)
-		}
-
-		// Expect AND keyword
-		if !p.isType(models.TokenTypeAnd) {
-			return nil, p.expectedError("AND")
-		}
-		p.advance() // Consume AND
-
-		// Parse upper bound - use parseStringConcatExpression to support complex expressions
-		upper, err := p.parseStringConcatExpression()
-		if err != nil {
-			return nil, goerrors.InvalidSyntaxError(
-				fmt.Sprintf("failed to parse BETWEEN upper bound: %v", err),
-				p.currentLocation(),
-				p.currentToken.Token.Value,
-			)
-		}
-
-		return &ast.BetweenExpression{
-			Expr:  left,
-			Lower: lower,
-			Upper: upper,
-			Not:   notPrefix,
-			Pos:   betweenPos,
-		}, nil
-	}
-
-	// Check for LIKE/ILIKE operator
-	if p.isType(models.TokenTypeLike) || strings.EqualFold(p.currentToken.Token.Value, "ILIKE") {
-		operator := p.currentToken.Token.Value
-		// Reject ILIKE in non-PostgreSQL dialects — it is a PostgreSQL extension.
-		if strings.EqualFold(operator, "ILIKE") &&
-			p.dialect != "" &&
-			p.dialect != string(keywords.DialectPostgreSQL) {
-			return nil, fmt.Errorf(
-				"ILIKE is a PostgreSQL-specific operator and is not supported in %s; "+
-					"use LIKE or LOWER() for case-insensitive matching", p.dialect,
-			)
-		}
-		p.advance() // Consume LIKE/ILIKE
-
-		// Parse pattern
-		pattern, err := p.parsePrimaryExpression()
-		if err != nil {
-			return nil, goerrors.InvalidSyntaxError(
-				fmt.Sprintf("failed to parse LIKE pattern: %v", err),
-				p.currentLocation(),
-				p.currentToken.Token.Value,
-			)
-		}
-
-		return &ast.BinaryExpression{
-			Left:     left,
-			Operator: operator,
-			Right:    pattern,
-			Not:      notPrefix,
-		}, nil
-	}
-
-	// Check for REGEXP/RLIKE operator (MySQL)
-	if strings.EqualFold(p.currentToken.Token.Value, "REGEXP") || strings.EqualFold(p.currentToken.Token.Value, "RLIKE") {
-		operator := strings.ToUpper(p.currentToken.Token.Value)
-		p.advance()
-		pattern, err := p.parsePrimaryExpression()
-		if err != nil {
-			return nil, goerrors.InvalidSyntaxError(
-				fmt.Sprintf("failed to parse REGEXP pattern: %v", err),
-				p.currentLocation(),
-				p.currentToken.Token.Value,
-			)
-		}
-		return &ast.BinaryExpression{
-			Left:     left,
-			Operator: operator,
-			Right:    pattern,
-			Not:      notPrefix,
-		}, nil
-	}
-
-	// Check for IN operator
-	if p.isType(models.TokenTypeIn) {
-		inPos := p.currentLocation()
-		p.advance() // Consume IN
-
-		// Expect opening parenthesis
-		if !p.isType(models.TokenTypeLParen) {
-			return nil, p.expectedError("(")
-		}
-		p.advance() // Consume (
-
-		// Check if this is a subquery (starts with SELECT or WITH)
-		if p.isType(models.TokenTypeSelect) || p.isType(models.TokenTypeWith) {
-			// Parse subquery
-			subquery, err := p.parseSubquery()
-			if err != nil {
-				return nil, goerrors.InvalidSyntaxError(
-					fmt.Sprintf("failed to parse IN subquery: %v", err),
-					p.currentLocation(),
-					p.currentToken.Token.Value,
-				)
-			}
-
-			// Expect closing parenthesis
-			if !p.isType(models.TokenTypeRParen) {
-				return nil, p.expectedError(")")
-			}
-			p.advance() // Consume )
-
-			return &ast.InExpression{
-				Expr:     left,
-				Subquery: subquery,
-				Not:      notPrefix,
-				Pos:      inPos,
-			}, nil
-		}
-
-		// Parse value list
-		var values []ast.Expression
-		for {
-			value, err := p.parseExpression()
-			if err != nil {
-				return nil, goerrors.InvalidSyntaxError(
-					fmt.Sprintf("failed to parse IN value: %v", err),
-					p.currentLocation(),
-					"",
-				)
-			}
-			values = append(values, value)
-
-			if p.isType(models.TokenTypeComma) {
-				p.advance() // Consume comma
-			} else if p.isType(models.TokenTypeRParen) {
-				break
-			} else {
-				return nil, p.expectedError(", or )")
-			}
-		}
-		p.advance() // Consume )
-
-		return &ast.InExpression{
-			Expr: left,
-			List: values,
-			Not:  notPrefix,
-			Pos:  inPos,
-		}, nil
-	}
-
-	// If NOT was consumed but no BETWEEN/LIKE/IN follows, we need to handle this case
-	// Put back the NOT by creating a NOT expression with left as the operand
-	if notPrefix {
-		return nil, goerrors.ExpectedTokenError(
-			"BETWEEN, LIKE, or IN",
-			"NOT",
-			p.currentLocation(),
-			"",
-		)
-	}
-
-	// Check for IS NULL / IS NOT NULL
-	if p.isType(models.TokenTypeIs) {
-		isPos := p.currentLocation()
-		p.advance() // Consume IS
-
-		isNot := false
-		if p.isType(models.TokenTypeNot) {
-			isNot = true
-			p.advance() // Consume NOT
-		}
-
-		if p.isType(models.TokenTypeNull) {
-			p.advance() // Consume NULL
-			return &ast.BinaryExpression{
-				Left:     left,
-				Operator: "IS NULL",
-				Right:    &ast.LiteralValue{Value: nil, Type: "null"},
-				Not:      isNot,
-				Pos:      isPos,
-			}, nil
-		}
-
-		return nil, p.expectedError("NULL")
-	}
-
-	// Check if this is a comparison binary expression (uses O(1) switch instead of O(n) isAnyType)
-	if p.isComparisonOperator() {
-		// Save the operator and its position
-		opPos := p.currentLocation()
-		operator := p.currentToken.Token.Value
-		p.advance()
-
-		// Check for ANY/ALL subquery operators (uses O(1) switch instead of O(n) isAnyType)
-		if p.isQuantifier() {
-			quantifier := p.currentToken.Token.Value
-			p.advance() // Consume ANY/ALL
-
-			// Expect opening parenthesis
-			if !p.isType(models.TokenTypeLParen) {
-				return nil, p.expectedError("(")
-			}
-			p.advance() // Consume (
-
-			// Parse subquery
-			subquery, err := p.parseSubquery()
-			if err != nil {
-				return nil, goerrors.InvalidSyntaxError(
-					fmt.Sprintf("failed to parse %s subquery: %v", quantifier, err),
-					p.currentLocation(),
-					"",
-				)
-			}
-
-			// Expect closing parenthesis
-			if !p.isType(models.TokenTypeRParen) {
-				return nil, p.expectedError(")")
-			}
-			p.advance() // Consume )
-
-			if quantifier == "ANY" {
-				return &ast.AnyExpression{
-					Expr:     left,
-					Operator: operator,
-					Subquery: subquery,
-				}, nil
-			}
-			return &ast.AllExpression{
-				Expr:     left,
-				Operator: operator,
-				Subquery: subquery,
-			}, nil
-		}
-
-		// Parse the right side of the expression
-		right, err := p.parsePrimaryExpression()
-		if err != nil {
-			return nil, err
-		}
-
-		// Create a binary expression
-		return &ast.BinaryExpression{
-			Left:     left,
-			Operator: operator,
-			Right:    right,
-			Pos:      opPos,
-		}, nil
 	}
 
 	return left, nil
@@ -660,812 +379,4 @@ func (p *Parser) isJSONOperator() bool {
 		return true
 	}
 	return false
-}
-
-// parsePrimaryExpression parses a primary expression (literals, identifiers, function calls)
-func (p *Parser) parsePrimaryExpression() (ast.Expression, error) {
-	// Handle unary minus/plus (-6, -3.14, +x)
-	if p.isType(models.TokenTypeMinus) || p.isType(models.TokenTypePlus) {
-		unaryPos := p.currentLocation()
-		var op ast.UnaryOperator
-		if p.isType(models.TokenTypeMinus) {
-			op = ast.Minus
-		} else {
-			op = ast.Plus
-		}
-		p.advance() // Consume -/+
-		expr, err := p.parsePrimaryExpression()
-		if err != nil {
-			return nil, err
-		}
-		return &ast.UnaryExpression{
-			Operator: op,
-			Expr:     expr,
-			Pos:      unaryPos,
-		}, nil
-	}
-
-	if p.isType(models.TokenTypeCase) {
-		// Handle CASE expressions (both simple and searched forms)
-		return p.parseCaseExpression()
-	}
-
-	if p.isType(models.TokenTypeCast) {
-		// Handle CAST(expr AS type) expressions
-		return p.parseCastExpression()
-	}
-
-	if p.isType(models.TokenTypeInterval) {
-		// Handle INTERVAL 'value' expressions
-		return p.parseIntervalExpression()
-	}
-
-	if p.isType(models.TokenTypeArray) {
-		// Handle ARRAY[...] or ARRAY(SELECT ...) constructor
-		return p.parseArrayConstructor()
-	}
-
-	// Handle MySQL VALUES(column) helper in ON DUPLICATE KEY UPDATE.
-	// VALUES is normally a DML keyword, but inside ON DUPLICATE KEY UPDATE it acts
-	// as a scalar function that returns the value that was attempted for insertion.
-	// e.g.: INSERT INTO t (id, name) VALUES (1, 'Alice') ON DUPLICATE KEY UPDATE name=VALUES(name)
-	if p.isType(models.TokenTypeValues) && p.peekToken().Token.Type == models.TokenTypeLParen {
-		valuesPos := p.currentLocation()
-		p.advance() // Consume VALUES
-		funcCall, err := p.parseFunctionCall("VALUES")
-		if err != nil {
-			return nil, err
-		}
-		if funcCall.Pos.IsZero() {
-			funcCall.Pos = valuesPos
-		}
-		return funcCall, nil
-	}
-
-	// Handle keywords that can be used as function names in MySQL (IF, REPLACE, etc.)
-	if (p.isType(models.TokenTypeIf) || p.isType(models.TokenTypeReplace)) && p.peekToken().Token.Type == models.TokenTypeLParen {
-		kwPos := p.currentLocation()
-		identName := p.currentToken.Token.Value
-		p.advance()
-		funcCall, err := p.parseFunctionCall(identName)
-		if err != nil {
-			return nil, err
-		}
-		if funcCall.Pos.IsZero() {
-			funcCall.Pos = kwPos
-		}
-		return funcCall, nil
-	}
-
-	if p.isType(models.TokenTypeIdentifier) || p.isType(models.TokenTypeDoubleQuotedString) || (p.dialect == string(keywords.DialectSQLServer) && p.isNonReservedKeyword()) {
-		// Handle identifiers and function calls
-		// Double-quoted strings are treated as identifiers in SQL (e.g., "column_name")
-		// Non-reserved keywords (TARGET, SOURCE, etc.) can also be used as identifiers
-		identPos := p.currentLocation()
-		identName := p.currentToken.Token.Value
-		p.advance()
-
-		// Check for function call (identifier followed by parentheses)
-		if p.isType(models.TokenTypeLParen) {
-			// This is a function call
-			funcCall, err := p.parseFunctionCall(identName)
-			if err != nil {
-				return nil, err
-			}
-			// Assign position of function name
-			if funcCall.Pos.IsZero() {
-				funcCall.Pos = identPos
-			}
-
-			// MySQL MATCH(...) AGAINST(...) full-text search
-			if strings.EqualFold(identName, "MATCH") && strings.EqualFold(p.currentToken.Token.Value, "AGAINST") {
-				return p.parseMatchAgainst(funcCall)
-			}
-
-			return funcCall, nil
-		}
-
-		// Handle regular identifier or qualified identifier (table.column or table.*)
-		ident := &ast.Identifier{Name: identName, Pos: identPos}
-
-		// Check for qualified identifier (table.column) or qualified asterisk (table.*)
-		if p.isType(models.TokenTypePeriod) {
-			p.advance() // Consume .
-			if p.isType(models.TokenTypeAsterisk) || p.isType(models.TokenTypeMul) {
-				// Handle table.* (qualified asterisk).
-				// Both TokenTypeAsterisk and TokenTypeMul represent '*'.
-				ident = &ast.Identifier{
-					Table: ident.Name,
-					Name:  "*",
-					Pos:   identPos,
-				}
-				p.advance()
-			} else if p.isIdentifier() {
-				// Handle table.column (qualified identifier)
-				ident = &ast.Identifier{
-					Table: ident.Name,
-					Name:  p.currentToken.Token.Value,
-					Pos:   identPos,
-				}
-				p.advance()
-			} else {
-				return nil, goerrors.InvalidSyntaxError(
-					"expected column name or * after table qualifier",
-					p.currentLocation(),
-					"Use table.column or table.* syntax",
-				)
-			}
-		}
-
-		// Check for array subscript or slice syntax: identifier[...]
-		// This handles: arr[1], arr[1][2], arr[1:3], arr[2:], arr[:5]
-		if p.isType(models.TokenTypeLBracket) {
-			return p.parseArrayAccessExpression(ident)
-		}
-
-		return ident, nil
-	}
-
-	if p.isType(models.TokenTypeAsterisk) || p.isType(models.TokenTypeMul) {
-		// Handle asterisk (e.g., in COUNT(*) or SELECT *).
-		// Both TokenTypeAsterisk and TokenTypeMul represent '*' from the tokenizer.
-		p.advance()
-		return &ast.Identifier{Name: "*"}, nil
-	}
-
-	if p.isStringLiteral() {
-		// Handle string literals
-		value := p.currentToken.Token.Value
-		p.advance()
-		return &ast.LiteralValue{Value: value, Type: "string"}, nil
-	}
-
-	if p.isNumericLiteral() {
-		// Handle numeric literals (int or float)
-		value := p.currentToken.Token.Value
-		litType := "int"
-		if strings.ContainsAny(value, ".eE") {
-			litType = "float"
-		}
-		p.advance()
-		return &ast.LiteralValue{Value: value, Type: litType}, nil
-	}
-
-	if p.isBooleanLiteral() {
-		// Handle boolean literals (uses O(1) switch instead of O(n) isAnyType)
-		value := p.currentToken.Token.Value
-		p.advance()
-		return &ast.LiteralValue{Value: value, Type: "bool"}, nil
-	}
-
-	if p.isType(models.TokenTypePlaceholder) {
-		// Handle SQL placeholders (e.g., $1, $2 for PostgreSQL; @param for SQL Server)
-		value := p.currentToken.Token.Value
-		p.advance()
-		return &ast.LiteralValue{Value: value, Type: "placeholder"}, nil
-	}
-
-	if p.isType(models.TokenTypeNull) {
-		// Handle NULL literal
-		p.advance()
-		return &ast.LiteralValue{Value: nil, Type: "null"}, nil
-	}
-
-	if p.isType(models.TokenTypeLParen) {
-		// Handle parenthesized expression or subquery
-		parenPos := p.currentLocation()
-		p.advance() // Consume (
-
-		// Check if this is a subquery (starts with SELECT or WITH)
-		if p.isType(models.TokenTypeSelect) || p.isType(models.TokenTypeWith) {
-			// Parse subquery
-			subquery, err := p.parseSubquery()
-			if err != nil {
-				return nil, goerrors.InvalidSyntaxError(
-					fmt.Sprintf("failed to parse subquery: %v", err),
-					p.currentLocation(),
-					"",
-				)
-			}
-			// Expect closing parenthesis
-			if !p.isType(models.TokenTypeRParen) {
-				return nil, p.expectedError(")")
-			}
-			p.advance() // Consume )
-			return &ast.SubqueryExpression{Subquery: subquery, Pos: parenPos}, nil
-		}
-
-		// Regular parenthesized expression - could be tuple (a, b, c) or single (expr)
-		expr, err := p.parseExpression()
-		if err != nil {
-			return nil, err
-		}
-
-		// Check if this is a tuple (has comma after first expression)
-		if p.isType(models.TokenTypeComma) {
-			// This is a tuple expression (col1, col2, ...)
-			tuple := ast.GetTupleExpression()
-			tuple.Expressions = append(tuple.Expressions, expr)
-
-			for p.isType(models.TokenTypeComma) {
-				p.advance() // Consume comma
-				nextExpr, err := p.parseExpression()
-				if err != nil {
-					return nil, err
-				}
-				tuple.Expressions = append(tuple.Expressions, nextExpr)
-			}
-
-			if !p.isType(models.TokenTypeRParen) {
-				return nil, p.expectedError(")")
-			}
-			p.advance() // Consume )
-			return tuple, nil
-		}
-
-		// Expect closing parenthesis for single expression
-		if !p.isType(models.TokenTypeRParen) {
-			return nil, p.expectedError(")")
-		}
-		p.advance() // Consume )
-
-		// Check for array subscript or slice on parenthesized expression
-		// This handles: (expr)[1], (SELECT arr)[2:3]
-		if p.isType(models.TokenTypeLBracket) {
-			return p.parseArrayAccessExpression(expr)
-		}
-
-		return expr, nil
-	}
-
-	if p.isType(models.TokenTypeExists) {
-		// Handle EXISTS (subquery)
-		p.advance() // Consume EXISTS
-
-		// Expect opening parenthesis
-		if !p.isType(models.TokenTypeLParen) {
-			return nil, p.expectedError("(")
-		}
-		p.advance() // Consume (
-
-		// Parse the subquery
-		subquery, err := p.parseSubquery()
-		if err != nil {
-			return nil, goerrors.InvalidSyntaxError(
-				fmt.Sprintf("failed to parse EXISTS subquery: %v", err),
-				p.currentLocation(),
-				"",
-			)
-		}
-
-		// Expect closing parenthesis
-		if !p.isType(models.TokenTypeRParen) {
-			return nil, p.expectedError(")")
-		}
-		p.advance() // Consume )
-
-		return &ast.ExistsExpression{Subquery: subquery}, nil
-	}
-
-	if p.isType(models.TokenTypeNot) {
-		// Handle NOT expression (NOT EXISTS, NOT boolean)
-		notPos := p.currentLocation()
-		p.advance() // Consume NOT
-
-		if p.isType(models.TokenTypeExists) {
-			// NOT EXISTS (subquery)
-			p.advance() // Consume EXISTS
-
-			if !p.isType(models.TokenTypeLParen) {
-				return nil, p.expectedError("(")
-			}
-			p.advance() // Consume (
-
-			subquery, err := p.parseSubquery()
-			if err != nil {
-				return nil, goerrors.InvalidSyntaxError(
-					fmt.Sprintf("failed to parse NOT EXISTS subquery: %v", err),
-					p.currentLocation(),
-					"",
-				)
-			}
-
-			if !p.isType(models.TokenTypeRParen) {
-				return nil, p.expectedError(")")
-			}
-			p.advance() // Consume )
-
-			// Return NOT EXISTS as a BinaryExpression with NOT flag
-			return &ast.BinaryExpression{
-				Left:     &ast.ExistsExpression{Subquery: subquery},
-				Operator: "NOT",
-				Right:    nil,
-				Not:      true,
-				Pos:      notPos,
-			}, nil
-		}
-
-		// NOT followed by other expression (boolean negation)
-		// Parse at comparison level for proper precedence: NOT (a > b), NOT active
-		expr, err := p.parseComparisonExpression()
-		if err != nil {
-			return nil, err
-		}
-		return &ast.UnaryExpression{
-			Operator: ast.Not,
-			Expr:     expr,
-			Pos:      notPos,
-		}, nil
-	}
-
-	return nil, goerrors.UnexpectedTokenError(
-		p.currentToken.Token.Type.String(),
-		p.currentToken.Token.Value,
-		p.currentLocation(),
-		"",
-	)
-}
-
-// parseCaseExpression parses a CASE expression (both simple and searched forms)
-//
-// Simple CASE: CASE expr WHEN value THEN result ... [ELSE result] END
-// Searched CASE: CASE WHEN condition THEN result ... [ELSE result] END
-func (p *Parser) parseCaseExpression() (*ast.CaseExpression, error) {
-	casePos := p.currentLocation()
-	p.advance() // Consume CASE
-
-	caseExpr := &ast.CaseExpression{
-		WhenClauses: make([]ast.WhenClause, 0),
-		Pos:         casePos,
-	}
-
-	// Check if this is a simple CASE (has a value expression) or searched CASE (no value)
-	// Simple CASE: CASE expr WHEN value THEN result
-	// Searched CASE: CASE WHEN condition THEN result
-	if !p.isType(models.TokenTypeWhen) {
-		// This is a simple CASE - parse the value expression
-		value, err := p.parseExpression()
-		if err != nil {
-			return nil, goerrors.InvalidSyntaxError(
-				fmt.Sprintf("failed to parse CASE value: %v", err),
-				p.currentLocation(),
-				"",
-			)
-		}
-		caseExpr.Value = value
-	}
-
-	// Parse WHEN clauses (at least one required)
-	for p.isType(models.TokenTypeWhen) {
-		whenPos := p.currentLocation()
-		p.advance() // Consume WHEN
-
-		// Parse the condition/value expression
-		condition, err := p.parseExpression()
-		if err != nil {
-			return nil, goerrors.InvalidSyntaxError(
-				fmt.Sprintf("failed to parse WHEN condition: %v", err),
-				p.currentLocation(),
-				"",
-			)
-		}
-
-		// Expect THEN keyword
-		if !p.isType(models.TokenTypeThen) {
-			return nil, p.expectedError("THEN")
-		}
-		p.advance() // Consume THEN
-
-		// Parse the result expression
-		result, err := p.parseExpression()
-		if err != nil {
-			return nil, goerrors.InvalidSyntaxError(
-				fmt.Sprintf("failed to parse THEN result: %v", err),
-				p.currentLocation(),
-				"",
-			)
-		}
-
-		caseExpr.WhenClauses = append(caseExpr.WhenClauses, ast.WhenClause{
-			Condition: condition,
-			Result:    result,
-			Pos:       whenPos,
-		})
-	}
-
-	// Check that we have at least one WHEN clause
-	if len(caseExpr.WhenClauses) == 0 {
-		return nil, goerrors.InvalidSyntaxError(
-			"CASE expression requires at least one WHEN clause",
-			p.currentLocation(),
-			"",
-		)
-	}
-
-	// Parse optional ELSE clause
-	if p.isType(models.TokenTypeElse) {
-		p.advance() // Consume ELSE
-
-		elseResult, err := p.parseExpression()
-		if err != nil {
-			return nil, goerrors.InvalidSyntaxError(
-				fmt.Sprintf("failed to parse ELSE result: %v", err),
-				p.currentLocation(),
-				"",
-			)
-		}
-		caseExpr.ElseClause = elseResult
-	}
-
-	// Expect END keyword
-	if !p.isType(models.TokenTypeEnd) {
-		return nil, p.expectedError("END")
-	}
-	p.advance() // Consume END
-
-	return caseExpr, nil
-}
-
-// parseCastExpression parses a CAST expression: CAST(expr AS type)
-//
-// Examples:
-//
-//	CAST(id AS VARCHAR)
-//	CAST(price AS DECIMAL(10,2))
-//	CAST(name AS VARCHAR(100))
-func (p *Parser) parseCastExpression() (*ast.CastExpression, error) {
-	// Consume CAST keyword
-	p.advance()
-
-	// Expect opening parenthesis
-	if !p.isType(models.TokenTypeLParen) {
-		return nil, p.expectedError("(")
-	}
-	p.advance() // Consume (
-
-	// Parse the expression to be cast
-	expr, err := p.parseExpression()
-	if err != nil {
-		return nil, err
-	}
-
-	// Expect AS keyword
-	if !p.isType(models.TokenTypeAs) {
-		return nil, p.expectedError("AS")
-	}
-	p.advance() // Consume AS
-
-	// Parse the target data type
-	// The type can be:
-	// - Simple type: VARCHAR, INT, DECIMAL, etc.
-	// - Type with precision: VARCHAR(100), DECIMAL(10,2), etc.
-	if !p.isType(models.TokenTypeIdentifier) {
-		return nil, p.expectedError("data type")
-	}
-
-	dataType := p.currentToken.Token.Value
-	p.advance() // Consume type name
-
-	// Check for type parameters (e.g., VARCHAR(100), DECIMAL(10,2))
-	if p.isType(models.TokenTypeLParen) {
-		p.advance() // Consume (
-
-		// Build the full type string including parameters
-		typeParams := "("
-		paramCount := 0
-
-		for !p.isType(models.TokenTypeRParen) {
-			if paramCount > 0 {
-				if !p.isType(models.TokenTypeComma) {
-					return nil, p.expectedError(", or )")
-				}
-				typeParams += p.currentToken.Token.Value
-				p.advance() // Consume comma
-			}
-
-			// Parse parameter (should be a number)
-			if !p.isNumericLiteral() && !p.isType(models.TokenTypeIdentifier) {
-				return nil, goerrors.InvalidSyntaxError(
-					"expected numeric type parameter",
-					p.currentLocation(),
-					"Use CAST(expr AS TYPE(precision[, scale]))",
-				)
-			}
-
-			typeParams += p.currentToken.Token.Value
-			p.advance()
-			paramCount++
-		}
-
-		typeParams += ")"
-		dataType += typeParams
-
-		if !p.isType(models.TokenTypeRParen) {
-			return nil, p.expectedError(")")
-		}
-		p.advance() // Consume )
-	}
-
-	// Expect closing parenthesis of CAST
-	if !p.isType(models.TokenTypeRParen) {
-		return nil, p.expectedError(")")
-	}
-	p.advance() // Consume )
-
-	return &ast.CastExpression{
-		Expr: expr,
-		Type: dataType,
-	}, nil
-}
-
-// parseIntervalExpression parses an INTERVAL expression: INTERVAL 'value'
-//
-// Examples:
-//
-//	INTERVAL '1 day'
-//	INTERVAL '2 hours'
-//	INTERVAL '1 year 2 months 3 days'
-//	INTERVAL '30 days'
-func (p *Parser) parseIntervalExpression() (*ast.IntervalExpression, error) {
-	// Consume INTERVAL keyword
-	p.advance()
-
-	// Support both PostgreSQL style: INTERVAL '1 day'
-	// and MySQL style: INTERVAL 30 DAY, INTERVAL 1 HOUR
-	if p.isStringLiteral() {
-		value := p.currentToken.Token.Value
-		p.advance()
-		return &ast.IntervalExpression{Value: value}, nil
-	}
-
-	// MySQL style: INTERVAL <number> <unit>
-	if p.isNumericLiteral() {
-		numStr := p.currentToken.Token.Value
-		p.advance()
-		// Expect a unit keyword (DAY, HOUR, MINUTE, SECOND, MONTH, YEAR, WEEK, etc.)
-		unit := strings.ToUpper(p.currentToken.Token.Value)
-		p.advance()
-		return &ast.IntervalExpression{Value: numStr + " " + unit}, nil
-	}
-
-	return nil, goerrors.InvalidSyntaxError(
-		"expected string literal or number after INTERVAL keyword",
-		p.currentLocation(),
-		"Use INTERVAL '1 day' or INTERVAL 1 DAY syntax",
-	)
-}
-
-// parseArrayConstructor parses PostgreSQL ARRAY constructor syntax.
-// Supports both ARRAY[...] (square bracket) and ARRAY(...) (subquery) forms.
-//
-// Examples:
-//
-//	ARRAY[1, 2, 3]                   - Array literal with square brackets
-//	ARRAY['a', 'b', 'c']             - String array
-//	ARRAY[x, y, z]                   - Array from expressions
-//	ARRAY(SELECT id FROM users)      - Array from subquery
-func (p *Parser) parseArrayConstructor() (*ast.ArrayConstructorExpression, error) {
-	p.advance() // Consume ARRAY
-
-	arrayExpr := ast.GetArrayConstructor()
-
-	// Check for square bracket syntax: ARRAY[...]
-	if p.isType(models.TokenTypeLBracket) {
-		p.advance() // Consume [
-
-		// Parse comma-separated list of expressions (can be empty)
-		if !p.isType(models.TokenTypeRBracket) {
-			for {
-				elem, err := p.parseExpression()
-				if err != nil {
-					return nil, err
-				}
-				arrayExpr.Elements = append(arrayExpr.Elements, elem)
-
-				if p.isType(models.TokenTypeComma) {
-					p.advance() // Consume comma
-				} else if p.isType(models.TokenTypeRBracket) {
-					break
-				} else {
-					return nil, p.expectedError(", or ]")
-				}
-			}
-		}
-
-		// Expect closing bracket
-		if !p.isType(models.TokenTypeRBracket) {
-			return nil, p.expectedError("]")
-		}
-		p.advance() // Consume ]
-
-		return arrayExpr, nil
-	}
-
-	// Check for parenthesis syntax: ARRAY(SELECT ...)
-	if p.isType(models.TokenTypeLParen) {
-		p.advance() // Consume (
-
-		// Expect a subquery
-		if !p.isType(models.TokenTypeSelect) && !p.isType(models.TokenTypeWith) {
-			return nil, p.expectedError("SELECT in ARRAY subquery")
-		}
-
-		subquery, err := p.parseSubquery()
-		if err != nil {
-			return nil, err
-		}
-
-		selectStmt, ok := subquery.(*ast.SelectStatement)
-		if !ok {
-			return nil, p.expectedError("SELECT statement in ARRAY subquery")
-		}
-		arrayExpr.Subquery = selectStmt
-
-		// Expect closing parenthesis
-		if !p.isType(models.TokenTypeRParen) {
-			return nil, p.expectedError(")")
-		}
-		p.advance() // Consume )
-
-		return arrayExpr, nil
-	}
-
-	return nil, p.expectedError("[ or (")
-}
-
-// parseSubquery parses a subquery (SELECT or WITH statement).
-// Expects current token to be SELECT or WITH.
-func (p *Parser) parseSubquery() (ast.Statement, error) {
-	if p.isType(models.TokenTypeWith) {
-		// WITH statement handles its own token consumption
-		return p.parseWithStatement()
-	}
-
-	if p.isType(models.TokenTypeSelect) {
-		p.advance() // Consume SELECT
-		return p.parseSelectWithSetOperations()
-	}
-
-	return nil, goerrors.ExpectedTokenError(
-		"SELECT or WITH",
-		p.currentToken.Token.Type.String(),
-		p.currentLocation(),
-		"",
-	)
-}
-
-// parseFunctionCall parses a function call with optional OVER clause for window functions.
-//
-// Examples:
-//
-//	COUNT(*) -> regular aggregate function
-//	ROW_NUMBER() OVER (ORDER BY id) -> window function with OVER clause
-
-// parseArrayAccessExpression parses array subscript and slice expressions.
-//
-// Supports:
-//   - Single subscript: arr[1]
-//   - Multi-dimensional subscript: arr[1][2][3]
-//   - Slice with both bounds: arr[1:3]
-//   - Slice from start: arr[:5]
-//   - Slice to end: arr[2:]
-//   - Full slice: arr[:]
-//
-// Examples:
-//
-//	tags[1]              -> ArraySubscriptExpression with single index
-//	matrix[2][3]         -> Nested ArraySubscriptExpression (multi-dimensional)
-//	arr[1:3]             -> ArraySliceExpression with start and end
-//	arr[2:]              -> ArraySliceExpression with start only
-//	arr[:5]              -> ArraySliceExpression with end only
-//	(SELECT arr)[1]      -> Array access on subquery result
-func (p *Parser) parseArrayAccessExpression(arrayExpr ast.Expression) (ast.Expression, error) {
-	// arrayExpr is the expression before the first '['
-	// We need to parse one or more '[...]' subscripts/slices
-
-	result := arrayExpr
-
-	// Loop to handle chained subscripts: arr[1][2][3]
-	for p.isType(models.TokenTypeLBracket) {
-		p.advance() // Consume [
-
-		// Check for empty brackets [] - this is an error
-		if p.isType(models.TokenTypeRBracket) {
-			return nil, goerrors.InvalidSyntaxError(
-				"empty array subscript [] is not allowed",
-				p.currentLocation(),
-				"Use arr[index] or arr[start:end] syntax",
-			)
-		}
-
-		// Check for slice starting with colon: arr[:end]
-		if p.isType(models.TokenTypeColon) {
-			p.advance() // Consume :
-
-			// Parse end expression (if not ']')
-			var endExpr ast.Expression
-			if !p.isType(models.TokenTypeRBracket) {
-				end, err := p.parseExpression()
-				if err != nil {
-					return nil, goerrors.InvalidSyntaxError(
-						fmt.Sprintf("failed to parse array slice end: %v", err),
-						p.currentLocation(),
-						"",
-					)
-				}
-				endExpr = end
-			}
-
-			// Expect closing bracket
-			if !p.isType(models.TokenTypeRBracket) {
-				return nil, p.expectedError("]")
-			}
-			p.advance() // Consume ]
-
-			// Create ArraySliceExpression with no start
-			sliceExpr := ast.GetArraySliceExpression()
-			sliceExpr.Array = result
-			sliceExpr.Start = nil
-			sliceExpr.End = endExpr
-			result = sliceExpr
-			continue
-		}
-
-		// Parse first expression (index or slice start)
-		firstExpr, err := p.parseExpression()
-		if err != nil {
-			return nil, goerrors.InvalidSyntaxError(
-				fmt.Sprintf("failed to parse array index/slice: %v", err),
-				p.currentLocation(),
-				"",
-			)
-		}
-
-		// Check if this is a slice (has colon) or subscript
-		if p.isType(models.TokenTypeColon) {
-			p.advance() // Consume :
-
-			// Parse end expression (if not ']')
-			var endExpr ast.Expression
-			if !p.isType(models.TokenTypeRBracket) {
-				end, err := p.parseExpression()
-				if err != nil {
-					return nil, goerrors.InvalidSyntaxError(
-						fmt.Sprintf("failed to parse array slice end: %v", err),
-						p.currentLocation(),
-						"",
-					)
-				}
-				endExpr = end
-			}
-
-			// Expect closing bracket
-			if !p.isType(models.TokenTypeRBracket) {
-				return nil, p.expectedError("]")
-			}
-			p.advance() // Consume ]
-
-			// Create ArraySliceExpression
-			sliceExpr := ast.GetArraySliceExpression()
-			sliceExpr.Array = result
-			sliceExpr.Start = firstExpr
-			sliceExpr.End = endExpr
-			result = sliceExpr
-		} else {
-			// This is a subscript, not a slice
-			// Expect closing bracket
-			if !p.isType(models.TokenTypeRBracket) {
-				return nil, p.expectedError("]")
-			}
-			p.advance() // Consume ]
-
-			// Create ArraySubscriptExpression with single index
-			subscriptExpr := ast.GetArraySubscriptExpression()
-			subscriptExpr.Array = result
-			subscriptExpr.Indices = append(subscriptExpr.Indices, firstExpr)
-			result = subscriptExpr
-		}
-	}
-
-	return result, nil
 }
